+++ resources/migrations/20220711142631-remove-workflow-licenses-table.down.sql
@@ -1,0 +1,7 @@
+CREATE TABLE if not exists workflow_licenses (
+  id serial NOT NULL PRIMARY KEY,
+  wfId integer DEFAULT NULL,
+  licId integer DEFAULT NULL,
+  CONSTRAINT workflow_licenses_ibfk_1 FOREIGN KEY (wfId) REFERENCES workflow (id),
+  CONSTRAINT workflow_licenses_ibfk_2 FOREIGN KEY (licId) REFERENCES license (id)
+);