+++ src/clj/rems/api/services/workflow.clj
@@ -1,58 +1,52 @@
 (ns rems.api.services.workflow
-  (:require [com.rpl.specter :refer [ALL transform]]
-            [rems.api.services.dependencies :as dependencies]
+  (:require [rems.api.services.dependencies :as dependencies]
             [rems.api.services.util :as util]
             [rems.db.applications :as applications]
             [rems.db.core :as db]
             [rems.db.form :as form]
+            [rems.db.licenses :as licenses]
             [rems.db.organizations :as organizations]
             [rems.db.users :as users]
-            [rems.db.workflow :as workflow]
-            [rems.json :as json]))
+            [rems.db.workflow :as workflow]))
 
 (defn invalid-forms-error [forms]
-  (let [invalid (seq (remove (comp form/get-form-template :form/id) forms))]
-    (when invalid
-      {:success false
-       :errors [{:type :invalid-form
-                 :forms invalid}]})))
+  (when-some [invalid (seq (remove (comp form/get-form-template :form/id) forms))]
+    {:success false
+     :errors [{:type :invalid-form
+               :forms invalid}]}))
 
 (defn invalid-users-error [userids]
-  (let [invalid (seq (remove users/user-exists? userids))]
-    (when invalid
-      {:success false
-       :errors [{:type :invalid-user
-                 :users invalid}]})))
+  (when-some [invalid (seq (remove users/user-exists? userids))]
+    {:success false
+     :errors [{:type :invalid-user
+               :users invalid}]}))
 
-(defn create-workflow! [{:keys [organization handlers forms] :as cmd}]
+(defn invalid-licenses-error [licenses]
+  (when-some [invalid (seq (remove (comp licenses/license-exists? :license/id) licenses))]
+    {:success false
+     :errors [{:type :invalid-license
+               :licenses invalid}]}))
+
+(defn create-workflow! [{:keys [organization handlers licenses forms] :as cmd}]
   (util/check-allowed-organization! organization)
   (or (invalid-users-error handlers)
       (invalid-forms-error forms)
-      (let [id (workflow/create-workflow! cmd)]
+      (invalid-licenses-error licenses)
+      (let [id (workflow/create-workflow! (update cmd :licenses #(map :license/id %)))]
         (dependencies/reset-cache!)
         {:success (not (nil? id))
          :id id})))
 
-(defn- unrich-workflow [workflow]
-  ;; TODO: keep handlers always in the same format, to avoid this conversion (we can ignore extra keys)
-  (if (get-in workflow [:workflow :handlers])
-    (update-in workflow [:workflow :handlers] #(map :userid %))
-    workflow))
-
-;; TODO: use rems.db.workflow/edit-workflow! and don't go directly to db fns, same for other fns
-(defn edit-workflow! [{:keys [id organization title handlers]}]
-  (let [workflow (unrich-workflow (workflow/get-workflow id))
-        workflow-body (cond-> (:workflow workflow)
-                        handlers (assoc :handlers handlers))]
+(defn edit-workflow! [{:keys [id organization handlers] :as cmd}]
+  (let [workflow (workflow/get-workflow id)]
     (util/check-allowed-organization! (:organization workflow))
     (when organization
       (util/check-allowed-organization! organization))
-    (db/edit-workflow! {:id id
-                        :title title
-                        :organization (:organization/id organization)
-                        :workflow (json/generate-string workflow-body)}))
-  (applications/reload-cache!)
-  {:success true})
+    (or (invalid-users-error handlers)
+        (do
+          (workflow/edit-workflow! (update cmd :licenses #(map :license/id %)))
+          (applications/reload-cache!)
+          {:success true}))))
 
 (defn set-workflow-enabled! [{:keys [id enabled]}]
   (util/check-allowed-organization! (:organization (workflow/get-workflow id)))
@@ -69,18 +63,21 @@
 
 ;; TODO more systematic joining for these needed. Now we just add the title for the UI
 (defn- enrich-workflow-form [item]
-  (select-keys (dependencies/enrich-dependency item) [:form/id :form/internal-name :form/external-title]))
+  (-> item
+      dependencies/enrich-dependency
+      (select-keys [:form/id :form/internal-name :form/external-title])))
 
-(defn- join-workflow-forms [workflow]
-  (update-in workflow [:workflow :forms] (partial mapv enrich-workflow-form)))
+(defn- enrich-workflow-license [item]
+  (-> item
+      licenses/join-license
+      organizations/join-organization))
 
 (defn- join-dependencies [workflow]
   (when workflow
-    (->> workflow
-         join-workflow-forms
-         organizations/join-organization
-         workflow/join-workflow-licenses
-         (transform [:licenses ALL] organizations/join-organization))))
+    (-> workflow
+        organizations/join-organization
+        (update-in [:workflow :forms] (partial map enrich-workflow-form))
+        (update-in [:workflow :licenses] (partial map enrich-workflow-license)))))
 
 (defn get-workflow [id]
   (->> (workflow/get-workflow id)
@@ -99,3 +96,4 @@
                            (get-in wf [:workflow :handlers]))
                          workflows)]
     (->> handlers distinct (sort-by :userid))))
+